feat: a whole new emacs config
Start a new "2023" version version of my emacs.d directory. It seems like a good
time to do this - I've been building on the same init.org file for 7 years, and
there's a fair bit of code that I don't use/need anymore, and as I upgrade emacs
and upgrade package versions I'm seeing some errors. Plus, the new
--init-directory setting in Emacs 29 makes it super easy to switch to another
emacs.d directory (which was possible before with chemacs and other tools, but
it's nice that it's builtin now).

This initial version contains a fair bit of default setup, org-mode and everyday
things. It doesn't have any real programming language support yet.

I'll try to be more organised with description and comments in this version - I
think it's worth the extra time investment as I'll inevitably forget what
something does or why it's there etc.