feat(emacs): add advice to fix org-agenda command display issue
Copying from the docstring:

Hacky advice to fix something I don't like about org-agenda.

When calling org-agenda, before the '*Agenda Commands*' buffer is switched to,
there's a manual call to (delete-other-windows). These windows are restored once
you've selected an agenda command, but I find it jarring to have my windows
temporarily disappear, and as the '*Agenda Commands*' buffer is quite small, it
isn't necessary to delete everything else on my screen.

Unfortunately there's no clean way to prevent this behaviour. It happens inside
(org-agenda-get-restriction-and-command) with an inlined call to
(delete-other-windows). This is a quite a long function, and isn't something
that I want to inline and redefine myself. So instead we have this hacky
dual-advice approach: this advice is intended for
org-agenda-get-restriction-and-command, which, when executed, will create
temporary advice to turn delete-other-windows into a no-op, and then restore
once the org function has exited.